requirements.txt
boto3>=1.26.0
requests>=2.28.0
pyyaml>=6.0
openai>=1.0.0
aws-cdk-lib>=2.100.0
constructs>=10.0.0
pyinstaller>=5.0.0
